+++ docs/source/api.rst
@@ -1,0 +1,12 @@
+API Reference
+=============
+
+.. automodule:: recipys.ingredients
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. autoclass:: recipys.ingredients.Ingredients
+   :members:
+   :undoc-members:
+   :show-inheritance:
