+++ tests/test_selectors.py
@@ -1,7 +1,7 @@
 import pytest
-from recipys.constants import Backend
+from recipies.constants import Backend
 
-from recipys.selector import (
+from recipies.selector import (
     Selector,
     all_outcomes,
     all_of,
